# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Corridor API
  description: |
    Corridor is a Business Orchestration Platform (BOAT) providing payment infrastructure,
    social payments, crowdfunding goals, and Earned Wage Access (EWA) services.
    
    ## Authentication
    Most endpoints require authentication via Bearer token or API key:
    - **Bearer Token**: For user-facing applications (login required)
    - **API Key**: For partner integrations (header: `X-API-Key`)
    
    ## Rate Limits
    - Auth endpoints: 10 requests/minute
    - Payment endpoints: 100 requests/minute  
    - Partner API: 1000 requests/minute (with API key)
  version: 1.0.0
  contact:
    email: people@corridormoney.net
    name: Corridor Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.corridormoney.net
    description: Production
  - url: https://api-sandbox.corridormoney.net
    description: Sandbox
  - url: http://localhost:8000
    description: Local Development
...
paths:
...
  /api/social/goals:
    get:
      tags: [Social Goals]
      summary: List user's crowdfunding goals
      operationId: getSocialGoals
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SocialGoal'
    post:
      tags: [Social Goals]
      summary: Create a new crowdfunding goal
      operationId: createSocialGoal
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Goal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocialGoal'
...
  /api/social/goals/contribute:
    post:
      tags: [Social Goals]
      summary: Contribute to a goal
      operationId: contributeToGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributeRequest'
      responses:
        '200':
          description: Contribution recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalContribution'
...
  /api/account/ewa/requests:
    get:
      tags: [EWA]
      summary: List EWA requests
      operationId: getEWARequests
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EWARequest'
    post:
      tags: [EWA]
      summary: Request an EWA withdrawal
      operationId: requestEWA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
      responses:
        '200':
          description: Request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EWARequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    SocialGoal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        target_amount:
          type: number
        current_amount:
          type: number
        currency:
          type: string
        product_link:
          type: string
        share_link:
          type: string
        status:
          type: string
          enum: [ACTIVE, COMPLETED, CANCELLED, EXPIRED]
        created_at:
          type: string
          format: date-time
    CreateGoalRequest:
      type: object
      required: [title, target_amount, currency]
      properties:
        title:
          type: string
        description:
          type: string
        target_amount:
          type: number
        currency:
          type: string
          enum: [USDC, USD, KES, NGN]
        product_link:
          type: string
    ContributeRequest:
      type: object
      required: [goal_id, amount, currency]
      properties:
        goal_id:
          type: string
          format: uuid
        contributor_name:
          type: string
          default: Anonymous
        amount:
          type: number
        currency:
          type: string
    GoalContribution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        goal_id:
          type: string
          format: uuid
        contributor_name:
          type: string
        amount:
          type: number
        currency:
          type: string
        transaction_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
    EWARequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        amount_requested:
          type: number
        status:
          type: string
          enum: [PENDING, APPROVED, DISBURSED, REJECTED]
        created_at:
          type: string
          format: date-time
...